Splitting NavierStokes.py into ChannelFlow2D.py and OrrSommerfeld.py
