+++ tests/test_demos.py
@@ -8,10 +8,13 @@ root = Path(__file__).parent.parent
 
 # Modules that live in examples/ but are not demos: imported by a demo rather
 # than run on their own. `runpy` puts examples/ on sys.path, which is what makes
-# those imports resolve. OrrSommerfeld_eigs supplies the eigenmode NavierStokes
-# seeds itself with; running it directly only reaches its argparse entry point,
-# which rejects pytest's argv and exits before computing anything.
-NOT_DEMOS = {"OrrSommerfeld_eigs"}
+# those imports resolve. ChannelFlow2D defines the channel solver and has no entry
+# point -- OrrSommerfeld and RayleighBenard drive it, and are collected here in
+# its place. Without this it would still be *collected*, and would pass having run
+# nothing at all. OrrSommerfeld_eigs supplies the eigenmode OrrSommerfeld seeds
+# itself with; running it directly only reaches its argparse entry point, which
+# rejects pytest's argv and exits before computing anything.
+NOT_DEMOS = {"OrrSommerfeld_eigs", "ChannelFlow2D"}
 
 _all_files = [f for f in root.glob("examples/*.py") if f.is_file()]
 _all_files = [
